Improve splash screen and split widgets, dialogs, windows and utils from dvc_interface (#198)
* split dvc_interface
* speed up splash screen
* create Widgets and Dialogs in ui module
* added utilities
* update setup and ui directory to be part of the package
---------
