+++ recipe/meta.yaml
@@ -11,7 +11,7 @@ build:
   preserve_egg_dir: False
 
   entry_points:
-    - idvc = idvc.dvc_interface:main
+    - idvc = idvc.idvc:main
 
   missing_dso_whitelist:
     - /lib64/libc.so.6            # [linux]
