+++ src/idvc/dvc_interface.py
@@ -21,20 +21,19 @@ from PySide2.QtCore import (QByteArray, QRegExp, QSettings, QSize, Qt,
                             QThreadPool)
 from PySide2.QtGui import QCloseEvent, QKeySequence, QRegExpValidator
 from PySide2.QtWidgets import (QAction, QCheckBox, QComboBox,
-                               QDialog, QDialogButtonBox, QDockWidget,
+                               QDockWidget,
                                QDoubleSpinBox, QFileDialog, QFormLayout,
                                QFrame, QGroupBox, QLabel, QLineEdit,
                                QMainWindow, QMessageBox,
                                QProgressDialog, QPushButton, QSpinBox,
-                               QStatusBar, QStyle, QTabWidget, QVBoxLayout,
+                               QTabWidget, QVBoxLayout,
                                QHBoxLayout, QSizePolicy,
                                QWidget)
 import time
 import numpy as np
 import math
 
-from matplotlib.backends.backend_qt5agg import FigureCanvasQTAgg as FigureCanvas
-from matplotlib.backends.backend_qt5agg import NavigationToolbar2QT as NavigationToolbar
+
 import matplotlib.pyplot as plt
 
 from functools import partial
@@ -88,61 +87,17 @@ from qdarkstyle.dark.palette import DarkPalette
 from qdarkstyle.light.palette import LightPalette
 
 from idvc import version as gui_version
-import multiprocessing
+
+from idvc.ui.widgets import *
+from idvc.ui.dialogs import *
+from idvc.ui.windows import *
+from idvc.utilities import *
 
 __version__ = gui_version.version
 
 import logging
 
-class PrintCallback(object):
-    '''Class to handle the emit call when no callback is provided'''
-    def emit(self, *args, **kwargs):
-        print (args, kwargs)
 
-def reduce_displ(raw_displ, min_size, max_size, pzero=False):
-    '''filter the diplacement vectors based on their size'''
-    offset = 6 # 6 in the case of the iDVC
-    
-    # sizes = []
-    # dmin = np.inf
-    # dmax = 0.
-    # for el in raw_displ:
-    #     size = 0
-    #     for i in range(3):
-    #         #calculate size of vector
-    #         size += el[i+offset]*el[i+offset]
-    #     size = np.sqrt(size)
-    #     if size > dmax:
-    #         dmax = size
-    #     if size < dmin:
-    #         dmin = size
-    #     sizes.append(size)
-    vec = np.asarray(raw_displ)[:,offset:offset+3]
-    if pzero:
-        vec -= np.asarray(raw_displ[0][offset:offset+3])
-
-    sizes = np.sqrt( np.sum( np.power(vec, 2), axis=1) )
-
-    dmin = sizes.min()
-    dmax = sizes.max()
-    if min_size is None and max_size is None:
-        displ = raw_displ
-    else:
-        displ = []
-        if pzero:
-            for i in range(len(raw_displ)):
-                size = sizes[i]
-                if size > min_size  and size < max_size :
-                    line = raw_displ[i]
-                    line[offset:offset+3] -= vec[0]
-                    displ.append(line)
-        else:
-            for i in range(len(raw_displ)):
-                size = sizes[i]
-                if size > min_size  and size < max_size :
-                    displ.append(raw_displ[i])
-    displ = np.asarray(displ)
-    return displ, dmin, dmax
 
 class MainWindow(QMainWindow):
     def __init__(self):
@@ -5484,842 +5439,5 @@ Please select the new location of the file, or move it back to where it was orig
         
 
 
-class SettingsWindow(QDialog):
-
-    def __init__(self, parent):
-        super(SettingsWindow, self).__init__(parent)
-
-        self.parent = parent
-
-        self.setWindowTitle("Settings")
-
-        self.dark_checkbox = QCheckBox("Dark Mode")
-
-        self.copy_files_checkbox = QCheckBox("Allow a copy of the image files to be stored. ")
-        self.vis_size_label = QLabel("Maximum downsampled image size (GB): ")
-        self.vis_size_entry = QDoubleSpinBox()
-
-        self.vis_size_entry.setMaximum(64.0)
-        self.vis_size_entry.setMinimum(0.01)
-        self.vis_size_entry.setSingleStep(0.01)
-
-        if self.parent.settings.value("vis_size") is not None:
-            self.vis_size_entry.setValue(float(self.parent.settings.value("vis_size")))
-
-        else:
-            self.vis_size_entry.setValue(1.0)
-
-
-        if self.parent.settings.value("dark_mode") is not None:
-            if self.parent.settings.value("dark_mode") == "true":
-                self.dark_checkbox.setChecked(True)
-            else:
-                self.dark_checkbox.setChecked(False)
-        else:
-            self.dark_checkbox.setChecked(True)
-
-        separator = QFrame()
-        separator.setFrameShape(QFrame.HLine)
-        separator.setFrameShadow(QFrame.Raised)
-        self.adv_settings_label = QLabel("Advanced")
-
-
-        self.gpu_label = QLabel("Please set the size of your GPU memory.")
-        self.gpu_size_label = QLabel("GPU Memory (GB): ")
-        self.gpu_size_entry = QDoubleSpinBox()
-
-
-        if self.parent.settings.value("gpu_size") is not None:
-            self.gpu_size_entry.setValue(float(self.parent.settings.value("gpu_size")))
-
-        else:
-            self.gpu_size_entry.setValue(1.0)
-
-        self.gpu_size_entry.setMaximum(64.0)
-        self.gpu_size_entry.setMinimum(0.00)
-        self.gpu_size_entry.setSingleStep(0.01)
-        self.gpu_checkbox = QCheckBox("Use GPU for volume render. (Recommended) ")
-        self.gpu_checkbox.setChecked(True) #gpu is default
-        if self.parent.settings.value("volume_mapper") == "cpu":
-            self.gpu_checkbox.setChecked(False)
-
-        if hasattr(self.parent, 'copy_files'):
-            self.copy_files_checkbox.setChecked(self.parent.copy_files)
-
-        self.omp_threads_entry = QSpinBox(self)
-        # default OMP_THREADS based on the number of cores available
-        n_cores = int(multiprocessing.cpu_count())
-        if n_cores > 16:
-            omp_threads = 16
-        elif n_cores > 8:
-            omp_threads = 8
-        elif n_cores > 4:
-            omp_threads = 4
-        elif n_cores > 2:
-            omp_threads = 2
-        else:
-            omp_threads = 1
-        
-        self.omp_threads_entry.setValue(omp_threads)
-        self.omp_threads_entry.setRange(1, n_cores)
-        self.omp_threads_entry.setSingleStep(1)
-        self.omp_threads_label = QLabel("OMP Threads: ")
-
-
-        self.layout = QVBoxLayout(self)
-        self.layout.addWidget(self.dark_checkbox)
-        self.layout.addWidget(self.copy_files_checkbox)
-        self.layout.addWidget(self.vis_size_label)
-        self.layout.addWidget(self.vis_size_entry)
-        self.layout.addWidget(separator)
-        self.layout.addWidget(self.adv_settings_label)
-        self.layout.addWidget(self.gpu_checkbox)
-        self.layout.addWidget(self.gpu_label)
-        self.layout.addWidget(self.gpu_size_label)
-        self.layout.addWidget(self.gpu_size_entry)
-
-        self.layout.addWidget(self.omp_threads_label)
-        self.layout.addWidget(self.omp_threads_entry)
-
-
-        self.buttons = QDialogButtonBox(
-           QDialogButtonBox.Save | QDialogButtonBox.Cancel,
-           Qt.Horizontal, self)
-        self.layout.addWidget(self.buttons)
-        self.buttons.accepted.connect(self.accept)
-        self.buttons.rejected.connect(self.quit)
-
-    def accept(self):
-        #self.parent.settings.setValue("settings_chosen", 1)
-        if self.dark_checkbox.isChecked():
-            self.parent.settings.setValue("dark_mode", True)
-        else:
-            self.parent.settings.setValue("dark_mode", False)
-        self.parent.SetAppStyle()
-
-        if self.copy_files_checkbox.isChecked():
-            self.parent.copy_files = 1 # save for this session
-            self.parent.settings.setValue("copy_files", 1) #save for next time we open app
-        else:
-            self.parent.copy_files = 0
-            self.parent.settings.setValue("copy_files", 0)
-
-        if self.gpu_checkbox.isChecked():
-            self.parent.settings.setValue("volume_mapper", "gpu")
-            self.parent.vis_widget_3D.volume_mapper = vtk.vtkSmartVolumeMapper()
-        else:
-            self.parent.settings.setValue("volume_mapper", "cpu")
-
-        self.parent.settings.setValue("gpu_size", float(self.gpu_size_entry.value()))
-        self.parent.settings.setValue("vis_size", float(self.vis_size_entry.value()))
-
-        if self.parent.settings.value("first_app_load") != "False":
-            self.parent.CreateSessionSelector("new window")
-            self.parent.settings.setValue("first_app_load", "False")
-            
-        self.parent.settings.setValue("omp_threads", str(self.omp_threads_entry.value()))
-        self.close()
-
-
-        #print(self.parent.settings.value("copy_files"))
-    def quit(self):
-        if self.parent.settings.value("first_app_load") != "False":
-            self.parent.CreateSessionSelector("new window")
-            self.parent.settings.setValue("first_app_load", "False")
-        self.close()
-        
-
-
-class SaveObjectWindow(QtWidgets.QWidget):
-    '''a window which will appear when saving a mask or pointcloud
-    '''
-        #self.copy_files_label = QLabel("Allow a copy of the image files to be stored: ")
-
-    def __init__(self, parent, object_type, save_only):
-        super().__init__()
-
-        #print(save_only)
-
-        self.parent = parent
-        self.object = object_type
-
-        if self.object == "mask":
-            self.setWindowTitle("Save Existing Mask")
-            self.label = QLabel("Save mask as:")
-        elif self.object == "pointcloud":
-            self.setWindowTitle("Save Existing Point Cloud")
-            self.label = QLabel("Save Point Cloud as:")
-
-
-        self.setWindowModality(QtCore.Qt.ApplicationModal)
-        #self.setInputMode(QtWidgets.QInputDialog.TextInput)
-
-        self.textbox = QLineEdit(self)
-        rx = QRegExp("[A-Za-z0-9]+")
-        validator = QRegExpValidator(rx, self.textbox) #need to check this
-        self.textbox.setValidator(validator)
-
-        self.save_button = QPushButton("Save")
-        self.quit_button = QPushButton("Discard")
-        self.save_button.clicked.connect(lambda: self.save(save_only))
-        self.quit_button.clicked.connect(self.quit)
-        
-        self.setWindowFlags(QtCore.Qt.WindowTitleHint )
-        #self.setCancelButtonText("New Session")
-        #self.setAttribute(Qt.WA_DeleteOnClose)
-        self.layout = QtWidgets.QFormLayout()
-        self.layout.addRow(self.label)
-        self.layout.addRow(self.textbox)
-        self.layout.addRow(self.save_button, self.quit_button)
-        self.setLayout(self.layout)
-
-
-    def save(self, save_only):
-        if self.object == "mask":
-            #Load Saved Session
-            #print("Write mask to file, then carry on")
-            filename = self.textbox.text() + ".mha"
-            shutil.copyfile(os.path.join(tempfile.tempdir, self.parent.mask_file), os.path.join(tempfile.tempdir, "Masks", filename))
-            self.parent.mask_parameters['masksList'].addItem(filename)
-            self.parent.mask_details[filename] = self.parent.mask_details['current']
-            #print(self.parent.mask_details)
-
-            self.parent.mask_parameters['loadButton'].setEnabled(True)
-            self.parent.mask_parameters['masksList'].setEnabled(True)
-
-
-            if not save_only:
-                #print("Not save only")
-                #would be better to move this elsewhere
-                self.parent.mask_worker = Worker(self.parent.extendMask)
-                self.parent.create_progress_window("Loading", "Loading Mask")
-                self.parent.mask_worker.signals.progress.connect(self.parent.progress)
-                self.parent.mask_worker.signals.finished.connect(self.parent.DisplayMask)
-                self.parent.threadpool.start(self.parent.mask_worker)
-                self.parent.progress_window.setValue(10)
-            
-        if self.object == "pointcloud":
-            filename = self.textbox.text() + ".roi"
-            shutil.copyfile(os.path.join(tempfile.tempdir, "latest_pointcloud.roi"), os.path.join(tempfile.tempdir, filename))
-
-            self.parent.pointcloud_parameters['loadButton'].setEnabled(True)
-            self.parent.pointcloud_parameters['pointcloudList'].setEnabled(True)
-            self.parent.pointcloud_parameters['pointcloudList'].addItem(filename)
-            self.parent.pointCloud_details[filename] = self.parent.pointCloud_details['latest_pointcloud.roi']
-            #print(self.parent.pointCloud_details)
-            #self.parent.createPointCloud()
-            if not save_only:
-                self.parent.PointCloudWorker("create")
-            
-
-        self.close()
-
-    def quit(self):
-        if self.object == "mask":
-            #would be better to move this elsewhere
-            self.parent.mask_worker = Worker(self.parent.extendMask)
-            self.parent.create_progress_window("Loading", "Loading Mask")
-            self.parent.mask_worker.signals.progress.connect(self.parent.progress)
-            self.parent.mask_worker.signals.finished.connect(self.parent.DisplayMask)
-            self.parent.threadpool.start(self.parent.mask_worker)
-            self.parent.progress_window.setValue(10)
-
-        if self.object == "pointcloud":
-            self.parent.PointCloudWorker("create")
-            #self.parent.createPointCloud()
-
-        self.close()
-
-class VisualisationWindow(QtWidgets.QMainWindow):
-    '''creates a window which will contain the VisualisationWidgets
-    '''
-    def __init__(self, parent):
-        super().__init__()
-        self.parent = parent
-        self.setMinimumSize(200,200)
-
-class VisualisationWidget(QtWidgets.QMainWindow):
-    '''creates a window with a QCILViewerWidget as the central widget
-    '''
-    def __init__(self, parent, viewer=viewer2D, interactorStyle=vlink.Linked2DInteractorStyle):
-        super().__init__()
-        self.parent = parent
-
-        self.e = ErrorObserver()
-        self.viewer = viewer
-        self.interactorStyle = interactorStyle
-        self.createEmptyFrame()
-        self.threadpool = QThreadPool()
-
-    def getViewer(self):
-        return self.frame.viewer
-
-    def getInteractor(self):
-        return self.getViewer().getInteractor()
-
-    def getInteractorStyle(self):
-        return self.getViewer().style
-
-    def getViewerType(self):
-        return self.viewer
 
         
-    def createEmptyFrame(self):
-        #print("empty")
-        self.frame = QCILViewerWidget(viewer=self.viewer, shape=(600,600), interactorStyle=self.interactorStyle)
-        self.setCentralWidget(self.frame)
-        self.image_file = [""]
-       
-    def displayImageData(self):
-        self.createEmptyFrame()
-        if self.viewer == viewer3D:
-            #set volume mapper according to user settings:
-            if self.parent.settings.value("volume_mapper") == "cpu":
-                self.frame.viewer.volume_mapper = vtk.vtkFixedPointVolumeRayCastMapper()
-                self.frame.viewer.volume.SetMapper(self.frame.viewer.volume_mapper)
-        else:
-            self.frame.viewer.setVisualisationDownsampling(self.parent.resample_rate)
-            self.frame.viewer.setDisplayUnsampledCoordinates(True)
-
-            vs_widgets = self.parent.visualisation_setting_widgets
-
-            vs_widgets['loaded_image_dims_value'].setVisible(True)
-            vs_widgets['loaded_image_dims_value'].setText(str(self.parent.unsampled_image_dimensions))
-
-            #print("resample rate: ", self.parent.resample_rate)
-
-            if self.parent.resample_rate != [1,1,1]:
-                vs_widgets['displayed_image_dims_value'].setVisible(True)
-                vs_widgets['displayed_image_dims_label'].setVisible(True)
-                #print("Disp image size ", [self.parent.ref_image_data.GetDimensions()[i] for i in range(3)])
-                vs_widgets['displayed_image_dims_value'].setText(str([round(self.parent.ref_image_data.GetDimensions()[i]) for i in range(3)]))
-                vs_widgets['coords_combobox'].setEnabled(True)
-                vs_widgets['coords_combobox'].setCurrentIndex(0)
-                vs_widgets['coords_warning_label'].setVisible(True)
-                vs_widgets['coords_info_label'].setVisible(True)
-
-            
-            else:
-                vs_widgets['displayed_image_dims_value'].setVisible(False)
-                vs_widgets['displayed_image_dims_label'].setVisible(False)
-                vs_widgets['coords_warning_label'].setVisible(False)
-                vs_widgets['coords_info_label'].setVisible(False)
-
-                vs_widgets['coords_combobox'].setEnabled(False)
-                vs_widgets['coords_combobox'].setCurrentIndex(0)
-
-        self.frame.viewer.setInput3DData(self.image_data)  
-        interactor = self.frame.viewer.getInteractor()
-
-
-        if hasattr(self.parent, 'orientation'):
-                orientation = self.parent.orientation
-        else:
-            orientation = self.frame.viewer.getSliceOrientation()
-        
-        if orientation == SLICE_ORIENTATION_XZ:
-            axis = 'y'
-        elif orientation == SLICE_ORIENTATION_YZ:
-            axis = 'x'
-        else:
-            axis = 'z'
-        interactor.SetKeyCode(axis)
-
-        if self.viewer == viewer2D:
-            self.frame.viewer.style.OnKeyPress(interactor, 'KeyPressEvent')
-            if self.parent.current_slice:
-                if self.parent.current_slice <= self.frame.viewer.img3D.GetExtent()[self.frame.viewer.getSliceOrientation()*2+1]:
-                    self.frame.viewer.displaySlice(self.parent.current_slice)
-
-
-        if self.viewer == viewer3D:
-            self.frame.viewer.style.OnKeyPress(interactor, 'KeyPressEvent')
-            # Depth peeling for volumes doesn't work as we would like when we have the vtk.vtkFixedPointVolumeRayCastMapper() instead of the vtk.vtkSmartVolumeMapper()
-            # self.frame.viewer.sliceActor.GetProperty().SetOpacity(0.99)
-            # self.frame.viewer.ren.SetUseDepthPeeling(True)
-            # self.frame.viewer.renWin.SetAlphaBitPlanes(True)
-            # self.frame.viewer.renWin.SetMultiSamples(False)
-            # self.frame.viewer.ren.UseDepthPeelingForVolumesOn()
-    
-            if self.parent.current_slice:
-                if self.parent.current_slice <= self.frame.viewer.img3D.GetExtent()[self.frame.viewer.getSliceOrientation()*2+1]:
-                    self.frame.viewer.style.SetActiveSlice(self.parent.current_slice)
-                    self.frame.viewer.style.UpdatePipeline()
-
-        # print("set input data for" + str(self.viewer))
-
-        if self.viewer == viewer2D:
-            self.PlaneClipper = cilPlaneClipper()
-            self.PlaneClipper.SetInteractorStyle(self.frame.viewer.style)
-
-
-    def setImageData(self, image_data):
-        self.image_data = image_data
-
-    def getImageData(self):
-        return self.image_data
-
-class GraphsWindow(QMainWindow):
-    '''creates a new window with graphs from results saved in the selected run folder.
-    '''
-    def __init__(self, parent=None):
-        super(GraphsWindow, self).__init__(parent)
-        self.setWindowTitle("Digital Volume Correlation Results")
-        DVCIcon = QtGui.QIcon()
-        DVCIcon.addFile("DVCIconSquare.png")
-
-        # Menu
-        self.menu = self.menuBar()
-        self.file_menu = self.menu.addMenu("File")
-        self.settings_menu = self.menu.addMenu("Settings")
-
-        displacement_setting_action = QAction("Show Displacement Relative to Reference Point 0", self)
-        displacement_setting_action.setCheckable(True)
-        displacement_setting_action.setChecked(False)
-        self.displacement_setting_action = displacement_setting_action
-
-        displacement_setting_action.triggered.connect(self.ReloadGraphs)
-        self.settings_menu.addAction(displacement_setting_action)
-
-
-        # Exit QAction
-        exit_action = QAction("Exit", self)
-        exit_action.setShortcut(QKeySequence.Quit)
-        exit_action.triggered.connect(self.close)
-        self.file_menu.addAction(exit_action)
-
-        #Tab positions:
-        self.setTabPosition(QtCore.Qt.AllDockWidgetAreas,QTabWidget.North)
-        self.setDockOptions(QMainWindow.ForceTabbedDocks)
-             
-        # Window dimensions
-        geometry = qApp.desktop().availableGeometry(self)
-
-        self.setGeometry(50,50, geometry.width()-100, geometry.height()-100)
-        #self.setFixedSize(geometry.width() * 0.6, geometry.height() * 0.8)
-
-    def SetResultsFolder(self, folder):
-        self.results_folder = folder
-        self.setWindowTitle("Digital Volume Correlation Results - {foldername}".format(foldername=os.path.basename(self.results_folder)))
-    
-    def ReloadGraphs(self):
-        self.DeleteAllWidgets()
-        self.CreateDockWidgets(displ_wrt_point0 = self.displacement_setting_action.isChecked())
-
-    def DeleteAllWidgets(self):
-         for current_dock in self.findChildren(QDockWidget):
-            current_dock.close()
-            del current_dock
-
-    def CreateDockWidgets(self, displ_wrt_point0 = False):
-        result_list=[]
-        #print(results_folder[0])
-        for folder in glob.glob(os.path.join(self.results_folder, "dvc_result_*")):
-            file_path = os.path.join(folder, os.path.basename(folder))
-            result = RunResults(file_path)
-            result_list.append(result)
-    
-            GraphWidget = SingleRunResultsWidget(self, result, displ_wrt_point0)
-            dock1 = QDockWidget(result.title,self)
-            dock1.setAllowedAreas(QtCore.Qt.RightDockWidgetArea)
-            dock1.setWidget(GraphWidget)
-            self.addDockWidget(QtCore.Qt.RightDockWidgetArea,dock1)
-    
-        prev = None
-
-        for current_dock in self.findChildren(QDockWidget):
-            if self.dockWidgetArea(current_dock) == QtCore.Qt.RightDockWidgetArea:
-                existing_widget = current_dock
-
-                if prev:
-                    self.tabifyDockWidget(prev,current_dock)
-                prev= current_dock
-        
-        SummaryTab = SummaryGraphsWidget(self, result_list)
-        dock = QDockWidget("Summary",self)
-        dock.setAllowedAreas(QtCore.Qt.RightDockWidgetArea)
-        dock.setWidget(SummaryTab)
-        self.addDockWidget(QtCore.Qt.RightDockWidgetArea,dock)
-        self.tabifyDockWidget(prev,dock)
-
-        dock.raise_() # makes summary panel the one that is open by default.
-
-class SingleRunResultsWidget(QtWidgets.QWidget):
-    '''creates a dockable widget which will display results from a single run of the DVC code
-    '''
-    def __init__(self, parent, plot_data, displ_wrt_point0 = False):
-        super().__init__()
-        self.parent = parent
-
-        self.figure = plt.figure()
-        self.canvas = FigureCanvas(self.figure)
-        self.toolbar = NavigationToolbar(self.canvas, self)
-
-        #Layout
-        self.layout = QtWidgets.QVBoxLayout()
-        self.layout.addWidget(self.toolbar)
-        self.layout.addWidget(self.canvas)
-        self.setLayout(self.layout)
-
-        self.CreateHistogram(plot_data, displ_wrt_point0)
-
-    def CreateHistogram(self, result, displ_wrt_point0):
-        displ = np.asarray(
-        PointCloudConverter.loadPointCloudFromCSV(result.disp_file,'\t')[:]
-        )
-        if displ_wrt_point0:
-            point0_disp = [displ[0][6],displ[0][7], displ[0][8]]
-            for count in range(len(displ)):
-                for i in range(3):
-                    displ[count][i+6] = displ[count][i+6] - point0_disp[i]
-
-        plot_data = [displ[:,i] for i in range(5, displ.shape[1])]
-
-        numGraphs = len(plot_data)
-        if numGraphs <= 3:
-            numRows = 1
-        else:
-            numRows = np.round(np.sqrt(numGraphs))
-        numColumns = np.ceil(numGraphs/numRows)
-
-        plotNum = 0
-        for array in plot_data:
-            plotNum = plotNum + 1
-            ax = self.figure.add_subplot(int(numRows), int(numColumns), int(plotNum))
-            ax.set_ylabel("")
-            #ax.set_xlabel(plot_titles[plotNum-1])
-            ax.set_title(result.plot_titles[plotNum-1])
-            ax.hist(array,20)
-
-        plt.tight_layout() # Provides proper spacing between figures
-
-        self.canvas.draw() 
-
-class SummaryGraphsWidget(QtWidgets.QWidget):
-    '''creates a dockable widget which will display results from all runs in a bulk run
-    '''
-    def __init__(self, parent, result_list, displ_wrt_point0 = False):
-        super().__init__()
-        self.parent = parent
-
-        #Layout
-        self.layout = QtWidgets.QGridLayout()
-        #self.layout.setSpacing(1)
-        self.layout.setAlignment(Qt.AlignTop)
-
-        widgetno=0
-
-        if len(result_list) >=1:
-            result = result_list[0] #These options were the same for all runs:
-
-            self.results_details_label = QLabel(self)
-            self.results_details_label.setText("Subvolume Geometry: {subvol_geom}\n\
-Maximum Displacement: {disp_max}\n\
-Degrees of Freedom: {num_srch_dof}\n\
-Objective Function: {obj_function}\n\
-Interpolation Type: {interp_type}\n\
-Rigid Body Offset: {rigid_trans}".format(subvol_geom=result.subvol_geom, \
-            disp_max=result.disp_max, num_srch_dof=str(result.num_srch_dof), obj_function=result.obj_function, \
-            interp_type=result.interp_type, rigid_trans=str(result.rigid_trans)))
-            self.layout.addWidget(self.results_details_label,widgetno,0,5,1)
-            self.results_details_label.setAlignment(Qt.AlignTop)        
-            widgetno+=1
-
-
-        self.label = QLabel(self)
-        self.label.setText("Select which variable would like to compare: ")
-        self.layout.addWidget(self.label,widgetno,1)
-
-        self.combo = QComboBox(self)
-        self.combo.addItems(result.plot_titles)
-        self.layout.addWidget(self.combo,widgetno,2)  
-        widgetno+=1
-
-        self.label1 = QLabel(self)
-        self.label1.setText("Select which parameter you would like to compare: ")
-        self.layout.addWidget(self.label1,widgetno,1)  
-        
-        self.combo1 = QComboBox(self)
-        self.param_list = ["All","Sampling Points in Subvolume", "Subvolume Size"]
-        self.combo1.addItems(self.param_list)
-        self.layout.addWidget(self.combo1,widgetno,2)
-        widgetno+=1
-
-        self.subvol_points=[]
-        self.subvol_sizes=[]
-
-        for result in result_list:
-            if result.subvol_points not in self.subvol_points:
-                self.subvol_points.append(result.subvol_points)
-            if result.subvol_size not in self.subvol_sizes:
-                self.subvol_sizes.append(result.subvol_size)
-        self.subvol_points.sort()
-        self.subvol_sizes.sort()
-
-        self.secondParamLabel = QLabel(self)
-        self.secondParamLabel.setText("Subvolume size:")
-        self.layout.addWidget(self.secondParamLabel,widgetno,1)
-        
-        self.secondParamCombo = QComboBox(self)
-        self.secondParamList = [str(i) for i in self.subvol_sizes]
-        self.secondParamCombo.addItems(self.secondParamList)
-        self.layout.addWidget(self.secondParamCombo,widgetno,2)
-        widgetno+=1
-
-        self.combo1.currentIndexChanged.connect(self.showSecondParam)
-        self.secondParamLabel.hide()
-        self.secondParamCombo.hide()
-
-        self.button = QtWidgets.QPushButton("Plot Histograms")
-        self.button.clicked.connect(partial(self.CreateHistogram,result_list, displ_wrt_point0))
-        self.layout.addWidget(self.button,widgetno,2)
-        widgetno+=1
-
-        self.figure = plt.figure()
-        
-        self.canvas = FigureCanvas(self.figure)
-        self.toolbar = NavigationToolbar(self.canvas, self)
-        self.layout.addWidget(self.toolbar,widgetno,0,1,3)
-        widgetno+=1
-        self.layout.addWidget(self.canvas,widgetno,0,3,3)
-        widgetno+=1
-
-        self.setLayout(self.layout)
-
-    def showSecondParam(self):
-        index = self.combo1.currentIndex()
-        if index ==0:
-            self.secondParamLabel.hide()
-            self.secondParamCombo.hide()
-
-        elif index == 1:
-            self.secondParamLabel.show()
-            self.secondParamCombo.show()
-            self.secondParamLabel.setText("Subvolume Size:")
-            self.secondParamCombo.clear()
-            self.secondParamCombo.addItems([str(i) for i in self.subvol_sizes])
-
-        elif index == 2:
-            self.secondParamLabel.show()
-            self.secondParamCombo.show()
-            self.secondParamLabel.setText("Points in Subvolume:")
-            self.secondParamCombo.clear()
-            newList = []
-            self.secondParamCombo.addItems([str(i) for i in self.subvol_points])   
-        
-    
-    def CreateHistogram(self, result_list, displ_wrt_point0):
-
-        self.figure.clear()
-
-        index = self.combo1.currentIndex()
-        
-        points_list = []
-
-        resultsToPlot= []
-
-        displacements = []
-
-        for result in result_list:
-            displ = np.asarray(
-            PointCloudConverter.loadPointCloudFromCSV(result.disp_file,'\t')[:]
-            )
-            if displ_wrt_point0:
-                point0_disp = [displ[0][6],displ[0][7], displ[0][8]]
-                for count in range(len(displ)):
-                    for i in range(3):
-                        displ[count][i+6] = displ[count][i+6] - point0_disp[i]
-
-            no_points = np.shape(displ[0])
-
-            if no_points not in points_list:
-                points_list.append(no_points)
-
-            if index == 1: # Points in subvolume is compared
-                if result.subvol_size != float(self.secondParamCombo.currentText()):
-                    pass
-
-            elif index ==2:
-                if result.subvol_points != float(self.secondParamCombo.currentText()):
-                    pass
-            
-            resultsToPlot.append(result)
-            displacements.append(displ)
-
-        points_list.sort()
-
-        if index ==0:
-            numRows = len(self.subvol_points)
-            numColumns = len(self.subvol_sizes)
-
-        else:
-            if len(resultsToPlot) <= 3:
-                numRows = 1
-            else:
-                numRows = np.round(np.sqrt(len(resultsToPlot)))
-            numColumns = np.ceil(len(resultsToPlot)/numRows)
-
-        plotNum = 0
-        for i, result in enumerate(resultsToPlot):
-            if index ==0:
-                row = self.subvol_points.index(result.subvol_points) + 1
-                column= self.subvol_sizes.index(result.subvol_size) + 1
-                plotNum = (row-1)*numColumns + column
-                ax = self.figure.add_subplot(numRows, numColumns, plotNum)
-                
-                if row ==1:
-                    ax.set_title("Subvolume Size:" + str(result.subvol_size) )
-                if column == 1:
-                    text = str(result.subvol_points) 
-                    ax.set_ylabel(text + " " + "Points in subvol")
-
-            else:
-                plotNum = plotNum + 1
-                ax = self.figure.add_subplot(numRows, numColumns, plotNum)
-    
-                if index ==1:
-                    text = str(result.subvol_points) 
-                if index ==2:
-                    text = str(result.subvol_size) 
-                ax.set_ylabel(text + " " + self.combo1.currentText())
-
-            plot_data = [displacements[i][:,k] for k in range(5, displacements[i].shape[1])]
-
-            #get variable to display graphs for:
-            ax.hist(plot_data[self.combo.currentIndex()], 20)
-
-        self.figure.suptitle(self.combo.currentText(),size ="large")
-
-        plt.tight_layout() # Provides proper spacing between figures
-        plt.subplots_adjust(top=0.88) # Means heading doesn't overlap with subplot titles
-        self.canvas.draw()
-        
-class RunResults(object):
-    def __init__(self, file_name):
-        
-        self.points = None
-
-        disp_file_name = file_name + ".disp"
-        stat_file_name = file_name + ".stat"
-
-        with open(stat_file_name,"r") as stat_file:
-            
-            count = 0
-            offset = 0
-            for line in stat_file:
-                if count == 9:
-                    if line.split('\t')[0] == "vol_endian":
-                        offset = 1
-
-                if count == 14 + offset:
-                    self.subvol_geom = str(line.split('\t')[1])
-                if count == 15 + offset:
-                    self.subvol_size = round(int(line.split('\t')[1]))
-                if count == 16 +offset:
-                    self.subvol_points = int(line.split('\t')[1])
-                if count == 20 + offset:
-                    self.disp_max = int(line.split('\t')[1])
-                if count == 21 + offset:
-                    self.num_srch_dof = int(line.split('\t')[1])
-                if count == 22 + offset:
-                    self.obj_function = str(line.split('\t')[1])
-                if count == 23 + offset:
-                    self.interp_type = str(line.split('\t')[1])
-                if count == 25 + offset:
-                    self.rigid_trans = [int(line.split('\t')[1]),int(line.split('\t')[2]), int(line.split('\t')[3])]
-                # if count == 26 + offset:
-                #     self.basin_radius = int(line.split('\t')[1])
-                # if count == 27 + offset:
-                #     self.subvol_aspect = [int(line.split('\t')[1]),int(line.split('\t')[2]), int(line.split('\t')[3])]
-                count+=1
-
-        plot_titles_dict = {
-            'objmin': "Objective Minimum", 'u': "Displacement in x", 'v':"Displacement in y", 'w':"Displacement in z",
-            'phi':"Change in phi",'the':"Change in theta", 'psi':"Change in psi"}
-
-        with open(disp_file_name) as f:
-            # first 4 columns are: n, x, y, z, status - we don't want these
-            self.plot_titles = f.readline().split()[5:]
-            self.plot_titles = [plot_titles_dict.get(text, text) for text in self.plot_titles]
-
-        
-        self.disp_file = disp_file_name
-
-        self.title =  str(self.subvol_points) + " Points in Subvolume," + " Subvolume Size: " + str(self.subvol_size)
-
-    def __str__(self):
-
-        a = "subvol_size {}".format(self.subvol_size)
-        n = "subvol_points {}".format(self.subvol_points)
-        return "RunResults:\n{}\n{}".format(a , n)
-
-def generateUIDockParameters(self, title): #copied from dvc_configurator.py
-    '''creates a dockable widget with a form layout group to add things to
-
-    basically you can add widget to the returned groupBoxFormLayout and paramsGroupBox
-    The returned dockWidget must be added with
-    self.addDockWidget(QtCore.Qt.RightDockWidgetArea, dockWidget)
-    '''
-    dockWidget = QDockWidget(self)
-    dockWidget.setWindowTitle(title)
-    dockWidgetContents = QWidget()
-
-
-    # Add vertical layout to dock contents
-    dockContentsVerticalLayout = QVBoxLayout(dockWidgetContents)
-    dockContentsVerticalLayout.setContentsMargins(0, 0, 0, 0)
-
-    # Create widget for dock contents
-    internalDockWidget = QWidget(dockWidgetContents)
-
-    # Add vertical layout to dock widget
-    internalWidgetVerticalLayout = QVBoxLayout(internalDockWidget)
-    internalWidgetVerticalLayout.setContentsMargins(0, 0, 0, 0)
-
-    # Add group box
-    paramsGroupBox = QGroupBox(internalDockWidget)
-
-
-    # Add form layout to group box
-    groupBoxFormLayout = QFormLayout(paramsGroupBox)
-    #groupBoxFormLayout.setFormAlignment(Qt.AlignCenter)
-
-    # Add elements to layout
-    internalWidgetVerticalLayout.addWidget(paramsGroupBox)
-    dockContentsVerticalLayout.addWidget(internalDockWidget)
-    dockWidget.setWidget(dockWidgetContents)
-
-    #        self.graphWidgetVL.addWidget(self.graphParamsGroupBox)
-    #        self.graphDockVL.addWidget(self.dockWidget)
-    #        self.pointCloudDockWidget.setWidget(self.pointCloudDockWidgetContents)
-    #
-    # self.addDockWidget(QtCore.Qt.RightDockWidgetArea, self.pointCloudDockWidget)
-    return (dockWidget, dockWidgetContents,
-            dockContentsVerticalLayout, internalDockWidget,
-            internalWidgetVerticalLayout, paramsGroupBox,
-            groupBoxFormLayout)
-
-
-def main():
-    app = QtWidgets.QApplication([])
-    
-    file_dir = os.path.dirname(__file__)
-    owl_file = os.path.join(file_dir, "DVCIconSquare.png")
-    owl = QtGui.QPixmap(owl_file)
-    splash = QtWidgets.QSplashScreen(owl)
-    splash.show()
-    
-    err = vtk.vtkFileOutputWindow()
-    err.SetFileName("../viewer.log")
-    vtk.vtkOutputWindow.SetInstance(err)
-
-    
-    window = MainWindow()
-    
-    window.show()
-    splash.finish(window)
-
-    sys.exit(app.exec_())
-
-if __name__ == "__main__":
-    main()
